Added the base of the dashboard and improved part 4
Added the base of the dashboard containing a sidebar where people can choose for an overview, an SIR model, country analysis and global analysis. Graphs from part 1 till 4 added to overview and country & global analysis. SIR model needs to be finished. Also improved plot_figures_country_complete in part 4 so that it accepted a start and end date for usage in the dashboard.